+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "include": [
     "src/**/*.ts",
+    "tests/**/*.ts"
   ],
   "compilerOptions": {
     "target": "es2022",
@@ -11,7 +12,7 @@
     "inlineSources": true,
     "strict": true,
     "outDir": "./extension/dist",
-    "rootDir": "./src",
+    "rootDir": "./",
     "moduleResolution": "bundler",
     "rewriteRelativeImportExtensions": true,
     "lib": [
